requirements.txt
numpy==1.24.3
Flask
scipy
pandas
matplotlib
seaborn
scikit-learn
opencv-python
Flask-Mail
flask-mysqldb
boto3
python-dotenv
python-decouple
nltk
spacy==2.3.5
tensorflow
fer
pyresparser
spacytextblob == 0.1.7
textblob == 0.15.3
fastapi==0.79.0
pandas==1.4.3
uvicorn==0.18.2
pydantic==1.9.1
pymongo~=4.2.0
motor~=3.0.0
sentence-transformers==2.2.2
pyjwt
<<<<<<< HEAD
pickle-mixin==1.0.2
joblib==0.16.0
openai
=======
>>>>>>> 2070008ba3f8d7c09ac13fc2c4f92b0dfd443131
